feat core: do not use requests_view if no one is interested in it
The only use of `Server::GetRequestsView` (hence the only use of `RequestsView`) is in `InspectRequests`, which is not even used in `MinimalServerComponentList`.

`RequestsView::DoJob` takes around ~1.2-1.5% in the perfflames i've got, not much, but still.

The proposed solution doesn't work correctly if `Server::GetRequestsView` is requested from outside of component constructor, but no one uses it anyway except for `InspectRequests`, and no one can, because `RequestsView` is private.


Tests: протестировано CI

Pull Request resolved: #144