feat server: implement standalone class for digest auth
The PR is dedicated to the AuthCheckerDigestBaseStandalone class, inherited from AuthCheckerDigestBase. It describes the logic of obtaining data (e.g. none, nonce_count) about the user and establishing data about the user.
There are also functions for adding nonces that are not tied to any user, and getting the creation time of these nonces. These two functions are necessary for the logic of the first authorization after the request without the Authorization header.
User information is stored in RCU map
Unnamed nonces are stored in the Expirable LRU Cache

Tests: протестировано CI
Pull Request resolved: #349

Co-authored-by: Konstantin Antoniadi <kos-antoniadi@yandex.ru>
Co-authored-by: Niyaz <nmingaripov@gmail.com>
Co-authored-by: Almaz Shagiev <68480736+bashkirian@users.noreply.github.com>
Co-authored-by: bashkirian <exaggeratethismotion@gmail.com>