feat core: decrease coroutines working set under normal service operation
This patch aims to reduce the coroutines working set: instead of using all the coroutines from pre-allocated pool, we still pre-allocate the pool, but try to reuse the coroutine as much as possible.
This should reduce the RES memory usage (and for some use cases -- drastically), because under normal service operation we de-virtualize much less mmap-ed coroutine stacks: say, if one had 5k coroutines in the pool and 200 alive coroutines max, before this patch we would minor-page-fault all the 5k stacks, with the patch -- only 200.

We didn't detect any noticeable performance changes with the patch, only if positive ones (due to working set reduction).