feat logs: PostgreSQL span names a little more informative
This is a proposal how to improve only PostgreSQL OpenTracing span names only. PR is related to https://github.com/userver-framework/userver/issues/678.
It sets names of spans according to their meaning and adds the first word of SQL query (or text before "as" if the first word is "with") to the name of span:

![image](https://github.com/user-attachments/assets/2caa20a7-f000-41d0-8650-c7a1b11a5dc4)

Regarding to Redis - it was decided to keep the span names for Redis unchanged, as otherwise, we would need to pass all arguments from Command to Reply and then convert them to strings, which would mean significant overhead for unclear reasons (name of command is the name of span already).

UPD: all the names were lowercased and only a set of predefined names is shown to avoid issues with comments and huge text blocks in `WITH` statements.

Tests: протестировано CI
Fixes: https://github.com/userver-framework/userver/issues/678
42b7ec21cb358bf4d7eaf8050cfd989e7d0de795

Pull Request resolved: https://github.com/userver-framework/userver/pull/681